Rename unit-name.ts to unit-label.ts
