+++ src/index.ts
@@ -3,7 +3,7 @@ import * as Quantity from "./quantity";
 import * as QuantityInfo from "./quantity-info";
 import * as Unit from "./unit";
 import * as Units from "./units";
-import * as UnitName from "./unit-name";
+import * as UnitLabel from "./unit-label";
 import * as UnitInfo from "./unit-info";
 import * as UnitDivide from "./unit-divide";
 import * as UnitTimes from "./unit-times";
@@ -14,7 +14,7 @@ export {
   QuantityInfo,
   Unit,
   Units,
-  UnitName,
+  UnitLabel,
   UnitInfo,
   UnitDivide,
   UnitTimes
